moved Exception class to utils
